Update LICENSE
this is an important development from my independent study of the pornography industry. stay tuned for big conjecture about the nature of digesting entertainment that promotes violence against women and is a leading GHG emitting past time given the server space and load (lulz) on the power grid. we will suggest that for one to offset their pornography addiction they will need to parse their personal library via gif generation and study of the things that make you go boom until you get over it and stop supporting the "bad" pornography market of revenge porn and non-consenting pornography (underage). we will add more to this but it is a good vein of what we have been discussing about the faults of cryptocurrency. cryptocurrency is a leading GHG emitter but so are banking practices. I've never had anyone even realize the two could be compared side by side and I would say the GHGs that result from conventional banking would be entire countries GHG emissions if one takes into account GDP somehow to normalize the system equations. like I said, more to come on these topics.