Update README.md
this derivation is tough because it can't be simplified to a fact and figure calculation that can be rationally debated if the reader wants to be pessimistic with regards to a point of interest of the writer. as Einstein would say, well... it's relative... you either think we are all going to be gutted by extremist anti capitalist terrorists and we need to shut the borders until the drug cartels pay their country club dues or you think living in a perpetual cyber sabbatical with digital cuddles and affordable healthcare is pretty neato and we should wait for marriage.